Bundle splitting (#266)
* Dynamic import demo

* use npm version of dynamic import plugin

* Add `import()` syntax for server-side rendering

I wrote a small package for this that replaces `import()` tokens with a
helper function that calls `require()`.
https://github.com/goto-bus-stop/dynamic-import-support

There's some other ways this could be done.

We could require that the app code supplies a polyfill for `import()`
for server side rendering if it needs it. It could use babel or
something to compile it first or just require `dynamic-import-support`
manually.

Or we could use `@std/esm` … https://github.com/standard-things/esm
That makes all of ES modules work. But then it'd be confusing that SSR
supports ES modules, while bundling doesn't—unless we add an ES modules
transform for browserify…

* Fix dynamic-import run order.

* exorcise dynamic bundles

* Fix ordering things more

* hacky take on minifying dynamic bundles

* dynamic import → split-require

* minify dynamic bundles using experimental tinyify api

* only tinyify dynamic bundles in prod

* use new tinyify

* use published split-require

* add dynamic bundle test

* disable bundle splitting for `bankai inspect`

* List split-require under ##Optimizations