hyperstream → hstream
hstream is a simpler version of hyperstream that uses the fast
`htmlparser2` parser directly. it does not support all the advanced
modifications that hyperstream can do, but it does support simply adding
tags to the `<head>`, which is all we need.

hstream is usually about 5x as fast as hyperstream.