Inject server rendered app at mount point.
Server rendering can reply with a `.selector` property indicating where
the rendered content should be mounted. It's optional; by default we
assume that we're rendering into the `<body>`.

This injects the server rendered content into the html _after_ all the
head tags are added, which should speed it up a bit for views with lots
of html. It detaches the initial template generation (the `head`
function) from injecting the server rendered response, so we can
implement custom index.html files (#343) later too.