Fix split-require integrity hashes
Previously we were telling split-require to generate a sha512 hash, but
then removing the source map from dynamic bundles afterwards,
invalidating the hash.

With this change we run exorcist as the very last part of the
split-require pipeline, by listening to the `split.pipeline` event. This
way split-require will generate hashes for the files _after_ source maps
have been externalised.

Also added integrity hashes to the prefetch links while debugging to
check whether they were the same; per
https://medium.com/reloading/preload-prefetch-and-priorities-in-chrome-776165961bbf#6dca
I'm not sure if this does anything or if prefetching at all might be
counterproductive.