Fix inlining critical css
https://github.com/choojs/bankai/pull/414 inserts the server rendered
app at the very end of the transform chain. The critical CSS inline
transform ran _before_ the server rendered app was inserted, so it did
not catch any selectors that were used in the app.

This patch moves the critical CSS transform to the end so that it works
on the entire server rendered app.